backend/services/order-management-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0                          # Sync/async HTTP for inter-service calls
confluent-kafka==2.6.1                 # MSK Serverless — publish order events
aws-msk-iam-sasl-signer-python==1.0.1  # MSK IAM SASL authentication
